[OPERATOR-500] Automation code struct refactor
