+++ test/integration_test/utils/utils.go
@@ -1,0 +1,42 @@
+package utils
+
+import (
+	"regexp"
+	"strings"
+
+	"github.com/hashicorp/go-version"
+)
+
+// MakeDNS1123Compatible will make the given string a valid DNS1123 name, which is the same
+// validation that Kubernetes uses for its object names.
+// Borrowed from
+// https://gitlab.com/gitlab-org/gitlab-runner/-/blob/0e2ae0001684f681ff901baa85e0d63ec7838568/executors/kubernetes/util.go#L268
+func MakeDNS1123Compatible(name string) string {
+	const (
+		DNS1123NameMaximumLength         = 63
+		DNS1123NotAllowedCharacters      = "[^-a-z0-9]"
+		DNS1123NotAllowedStartCharacters = "^[^a-z0-9]+"
+	)
+
+	name = strings.ToLower(name)
+
+	nameNotAllowedChars := regexp.MustCompile(DNS1123NotAllowedCharacters)
+	name = nameNotAllowedChars.ReplaceAllString(name, "")
+
+	nameNotAllowedStartChars := regexp.MustCompile(DNS1123NotAllowedStartCharacters)
+	name = nameNotAllowedStartChars.ReplaceAllString(name, "")
+
+	if len(name) > DNS1123NameMaximumLength {
+		name = name[0:DNS1123NameMaximumLength]
+	}
+
+	return name
+}
+
+// GetPxVersionFromSpecGenURL gets the px version to install or upgrade,
+// e.g. return version 2.9 for https://edge-install.portworx.com/2.9
+func GetPxVersionFromSpecGenURL(url string) *version.Version {
+	splitURL := strings.Split(url, "/")
+	v, _ := version.NewVersion(splitURL[len(splitURL)-1])
+	return v
+}
